pyproject.toml
[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[project]
name = "name-of-the-machine"
version = "0.0.1"
authors = [
    {name="Machina Ratiocinatrix", email="machina.ratio@gmail.com"},
    {name="Alexander Fedotov", email="alex.fedotov@aol.com"}
]
description = "Description of the Machine"
readme = "README.md"
requires-python = ">=3.10"
classifiers=[
        "Programming Language :: Python :: 3",
        "License :: OSI Approved :: MIT License",
        "Operating System :: OS Independent",
]
keywords = ["name-of-the-machine"]
dependencies = [
    "opehaina >= 0.0.5",
    "PyGithub >= 2.6.1",
    "PyYAML >= 6.0.2",
    "urllib3 >= 2.0.4",
    "requests >= 2.32.3",
    "click >= 8.3.0"
]


[project.optional-dependencies]
gemini = ["castor-pollux >= 0.1.4"]
anthropic = ["electroid >= 0.0.19"]
all = ["castor-pollux >= 0.1.4", "electroid >= 0.0.19"]

[project.scripts]
name-of-the-machine = "name_of_the_machine.cli:run"

[project.urls]
"Homepage" = "https://github.com/name-of-the-machine/name-of-the-machine"

[dependency-groups]
dev = ["pytest>=8.0.0"]